@@ -17,13 +17,13 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
-        "symfony/config": "^4.4 || ^5.1",
-        "symfony/dependency-injection": "^4.4 || ^5.1",
-        "symfony/form": "^4.4 || ^5.1",
-        "symfony/http-foundation": "^4.4 || ^5.1",
-        "symfony/http-kernel": "^4.4 || ^5.1",
-        "symfony/validator": "^4.4 || ^5.1",
+        "php": ">=8.0.2",
+        "symfony/config": "^5.4 || ^6.0",
+        "symfony/dependency-injection": "^5.4 || ^6.0",
+        "symfony/form": "^5.4 || ^6.0",
+        "symfony/http-foundation": "^5.4 || ^6.0",
+        "symfony/http-kernel": "^5.4 || ^6.0",
+        "symfony/validator": "^5.4 || ^6.0",
         "twig/twig": "~2.4 || ~3.0"
     },
     "require-dev": {
@@ -35,8 +35,8 @@
         "phpspec/phpspec": "^7.1",
         "phpstan/phpstan": "~0.12.99",
         "sebastian/comparator": "~4.0.6",
-        "symfony/twig-bridge": "^4.4 || ^5.1",
-        "symfony/var-dumper": "^4.4 || ^5.1"
+        "symfony/twig-bridge": "^5.4 || ^6.0",
+        "symfony/var-dumper": "^5.4 || ^6.0"
     },
     "minimum-stability": "stable",
     "autoload": {
